Time: 3 ms (11.11%) | Memory: 13.9 MB (12.96%) - LeetSync

+++ 1071-binary-prefix-divisible-by-5/binary-prefix-divisible-by-5.c
@@ -4,7 +4,7 @@
 bool* prefixesDivBy5(int* nums, int numsSize, int* returnSize) {
     bool* arr = (bool*)malloc(numsSize * sizeof(bool));
     int index = 0;
-    int count = 0;
+    unsigned long long count = 0;
     for(int i=0 ; i<numsSize ; i++){
         count = (count*2+nums[i])%5;
         if(count%5==0){
